refactor(appserver): isolate config handlers
